+++ src/main/kotlin/icu/windea/pls/localisation/intentions/CopyLocalisationPlainTextIntention.kt
@@ -1,6 +1,7 @@
 package icu.windea.pls.localisation.intentions
 
 import com.intellij.codeInsight.intention.*
+import com.intellij.codeInsight.intention.preview.*
 import com.intellij.openapi.editor.*
 import com.intellij.openapi.ide.*
 import com.intellij.openapi.project.*
@@ -37,5 +38,9 @@ class CopyLocalisationPlainTextIntention : IntentionAction {
 		val element = originalElement.parentOfType<ParadoxLocalisationProperty>() ?: return
 		val text = ParadoxLocalisationTextExtractor.extract(element)
 		CopyPasteManager.getInstance().setContents(StringSelection(text))
+	}
+	
+	override fun generatePreview(project: Project, editor: Editor, file: PsiFile): IntentionPreviewInfo {
+		return IntentionPreviewInfo.EMPTY
 	}
 }