+++ src/main/kotlin/icu/windea/pls/localisation/intentions/CopyLocalisationRichTextIntention.kt
@@ -1,6 +1,7 @@
 package icu.windea.pls.localisation.intentions
 
 import com.intellij.codeInsight.intention.*
+import com.intellij.codeInsight.intention.preview.*
 import com.intellij.openapi.editor.*
 import com.intellij.openapi.ide.*
 import com.intellij.openapi.project.*
@@ -37,6 +38,10 @@ class CopyLocalisationRichTextIntention : IntentionAction {
 		val element = originalElement.parentOfType<ParadoxLocalisationProperty>() ?: return
 		val text = ParadoxLocalisationTextRenderer.render(element)
 		CopyPasteManager.getInstance().setContents(StringSelection(text))
+	}
+	
+	override fun generatePreview(project: Project, editor: Editor, file: PsiFile): IntentionPreviewInfo {
+		return IntentionPreviewInfo.EMPTY
 	}
 }